etl/requirements.txt
psycopg2-binary==2.9.9   # External DB (Postgres)
pymssql==2.3.1            # Internal DB (MSSQL / SQL Server)
